src/test/java/com/sbtraining/recipe_project/services/RecipeServiceImplTest.java
package com.sbtraining.recipe_project.services;

import com.sbtraining.recipe_project.model.Recipe;
import com.sbtraining.recipe_project.repositories.RecipeRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.HashSet;
import java.util.Set;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

class RecipeServiceImplTest {

    RecipeServiceImpl recipeService;

    @Mock
    RecipeRepository recipeRepository;

    @BeforeEach
    void setUp() throws Exception {
        MockitoAnnotations.initMocks(this);
        recipeService = new RecipeServiceImpl(recipeRepository);
    }

    @Test
    void getAllRecipes() throws Exception {
        Recipe recipe1 = Recipe.builder().id(1L).build();
        Set<Recipe> recipesData = new HashSet<>();
        recipesData.add(recipe1);
        when(recipeRepository.findAll()).thenReturn(recipesData);
        Set<Recipe> recipeSet = recipeService.getAllRecipes();
        assertEquals(recipeSet.size(), 1);

        // BDD test
        verify(recipeRepository, times(1)).findAll();
    }
}